+++ backend/src/main/java/org/usyj/makgora/request/vote/VoteParticipateRequest.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter @Setter
 public class VoteParticipateRequest {
     // 사용자가 선택한 choice_id
-    private Integer choiceId;
+    private Long choiceId;
 
     // 베팅 포인트
     private Integer points;
